basic refactor. Still need to clean up tests, paths, cyclic dependencies and proper contributions.
