+++ io/fasta/data/base.fasta
