+++ io/fasta/fasta.go
