+++ io/fasta/fasta_test.go
