+++ io/genbank/genbank_test.go
@@ -1,0 +1,335 @@
+package genbank
+
+import (
+	"fmt"
+	"io/ioutil"
+	"os"
+	"path/filepath"
+	"strings"
+	"testing"
+
+	"github.com/TimothyStiles/poly"
+	"github.com/TimothyStiles/poly/io/json"
+	"github.com/google/go-cmp/cmp"
+	"github.com/google/go-cmp/cmp/cmpopts"
+)
+
+/******************************************************************************
+
+Gbk/gb/genbank related benchmarks begin here.
+
+******************************************************************************/
+
+func ExampleReadGbk() {
+	sequence := ReadGbk("../../data/puc19.gbk")
+	fmt.Println(sequence.Meta.Locus.ModificationDate)
+	// Output: 22-OCT-2019
+}
+
+func ExampleParseGbk() {
+	file, _ := ioutil.ReadFile("../../data/puc19.gbk")
+	sequence := ParseGbk(file)
+
+	fmt.Println(sequence.Meta.Locus.ModificationDate)
+	// Output: 22-OCT-2019
+}
+
+func ExampleBuildGbk() {
+	sequence := ReadGbk("../../data/puc19.gbk")
+	gbkBytes := BuildGbk(sequence)
+	testSequence := ParseGbk(gbkBytes)
+
+	fmt.Println(testSequence.Meta.Locus.ModificationDate)
+	// Output: 22-OCT-2019
+}
+
+func ExampleWriteGbk() {
+	tmpDataDir, err := ioutil.TempDir("", "data-*")
+	if err != nil {
+		fmt.Println(err.Error())
+	}
+	defer os.RemoveAll(tmpDataDir)
+
+	sequence := ReadGbk("../../data/puc19.gbk")
+
+	tmpGbkFilePath := filepath.Join(tmpDataDir, "puc19.gbk")
+	WriteGbk(sequence, tmpGbkFilePath)
+
+	testSequence := ReadGbk(tmpGbkFilePath)
+
+	fmt.Println(testSequence.Meta.Locus.ModificationDate)
+	// Output: 22-OCT-2019
+}
+
+func TestGbkIO(t *testing.T) {
+	tmpDataDir, err := ioutil.TempDir("", "data-*")
+	if err != nil {
+		t.Error(err)
+	}
+	defer os.RemoveAll(tmpDataDir)
+
+	gbk := ReadGbk("../../data/puc19.gbk")
+
+	tmpGbkFilePath := filepath.Join(tmpDataDir, "puc19.gbk")
+	WriteGbk(gbk, tmpGbkFilePath)
+
+	writeTestGbk := ReadGbk(tmpGbkFilePath)
+	if diff := cmp.Diff(gbk, writeTestGbk, cmpopts.IgnoreFields(poly.Feature{}, "ParentSequence")); diff != "" {
+		t.Errorf("Parsing the output of BuildGbk() does not produce the same output as parsing the original file read with ReadGbk(). Got this diff:\n%s", diff)
+	}
+
+	// Test multiline Genbank features
+	pichia := ReadGbk("../../data/pichia_chr1_head.gb")
+	var multilineOutput string
+	for _, feature := range pichia.Features {
+		multilineOutput = feature.GbkLocationString
+	}
+	if multilineOutput != "join(<459260..459456,459556..459637,459685..459739,459810..>460126)" {
+		t.Errorf("Failed to parse multiline genbank feature string")
+	}
+}
+
+func TestGbkLocationStringBuilder(t *testing.T) {
+	tmpDataDir, err := ioutil.TempDir("", "data-*")
+	if err != nil {
+		t.Error(err)
+	}
+	defer os.RemoveAll(tmpDataDir)
+
+	scrubbedGbk := ReadGbk("../../data/sample.gbk")
+
+	// removing gbkLocationString from features to allow testing for gbkLocationBuilder
+	for featureIndex := range scrubbedGbk.Features {
+		scrubbedGbk.Features[featureIndex].GbkLocationString = ""
+	}
+
+	tmpGbkFilePath := filepath.Join(tmpDataDir, "../../sample.gbk")
+	WriteGbk(scrubbedGbk, tmpGbkFilePath)
+
+	testInputGbk := ReadGbk("../../data/sample.gbk")
+	testOutputGbk := ReadGbk(tmpGbkFilePath)
+
+	if diff := cmp.Diff(testInputGbk, testOutputGbk, cmpopts.IgnoreFields(poly.Feature{}, "ParentSequence")); diff != "" {
+		t.Errorf("Issue with partial location building. Parsing the output of BuildGbk() does not produce the same output as parsing the original file read with ReadGbk(). Got this diff:\n%s", diff)
+	}
+}
+
+func TestGbLocationStringBuilder(t *testing.T) {
+	tmpDataDir, err := ioutil.TempDir("", "data-*")
+	if err != nil {
+		t.Error(err)
+	}
+	defer os.RemoveAll(tmpDataDir)
+
+	scrubbedGb := ReadGbk("../../data/t4_intron.gb")
+
+	// removing gbkLocationString from features to allow testing for gbkLocationBuilder
+	for featureIndex := range scrubbedGb.Features {
+		scrubbedGb.Features[featureIndex].GbkLocationString = ""
+	}
+
+	tmpGbFilePath := filepath.Join(tmpDataDir, "t4_intron_test.gb")
+	WriteGbk(scrubbedGb, tmpGbFilePath)
+
+	testInputGb := ReadGbk("../../data/t4_intron.gb")
+	testOutputGb := ReadGbk(tmpGbFilePath)
+
+	if diff := cmp.Diff(testInputGb, testOutputGb, cmpopts.IgnoreFields(poly.Feature{}, "ParentSequence")); diff != "" {
+		t.Errorf("Issue with either Join or complement location building. Parsing the output of BuildGbk() does not produce the same output as parsing the original file read with ReadGbk(). Got this diff:\n%s", diff)
+	}
+}
+
+func TestPartialLocationParseRegression(t *testing.T) {
+	gbk := ReadGbk("../../data/sample.gbk")
+
+	for _, feature := range gbk.Features {
+		if feature.GbkLocationString == "687..3158>" && (feature.SequenceLocation.Start != 686 || feature.SequenceLocation.End != 3158) {
+			t.Errorf("Partial location for three prime location parsing has failed. Parsing the output of BuildGbk() does not produce the same output as parsing the original file read with ReadGbk()")
+		} else if feature.GbkLocationString == "<1..206" && (feature.SequenceLocation.Start != 0 || feature.SequenceLocation.End != 206) {
+			t.Errorf("Partial location for five prime location parsing has failed. Parsing the output of BuildGbk() does not produce the same output as parsing the original file read with ReadGbk().")
+		}
+	}
+}
+
+func TestLocusParseRegression(t *testing.T) {
+	gbk := ReadGbk("../../data/puc19.gbk").Meta.Locus
+	json := json.ReadJSON("../../data/puc19static.json").Meta.Locus
+
+	if diff := cmp.Diff(gbk, json, cmpopts.IgnoreFields(poly.Feature{}, "ParentSequence")); diff != "" {
+		t.Errorf("The meta parser has changed behaviour. Got this diff:\n%s", diff)
+	}
+}
+
+func TestSnapgeneGenbankRegression(t *testing.T) {
+	snapgene := ReadGbk("../../data/puc19_snapgene.gb")
+
+	if snapgene.Sequence == "" {
+		t.Errorf("Parsing snapgene returned an empty string")
+	}
+}
+
+func TestGetSequenceMethods(t *testing.T) {
+
+	gbk := ReadGbk("../../data/t4_intron.gb")
+
+	// Check to see if GetSequence method works on Annotated struct
+	if gbk.GetSequence() != gbk.Sequence {
+		t.Errorf(" Sequence GetSequence method has failed'. Got this:\n%s instead of \n%s", gbk.GetSequence(), gbk.Sequence)
+	}
+
+	// Check to see if GetSequence method works on Features struct
+	feature := gbk.Features[1].GetSequence()
+	seq := "atgagattacaacgccagagcatcaaagattcagaagttagaggtaaatggtattttaatatcatcggtaaagattctgaacttgttgaaaaagctgaacatcttttacgtgatatgggatgggaagatgaatgcgatggatgtcctctttatgaagacggagaaagcgcaggattttggatttaccattctgacgtcgagcagtttaaagctgattggaaaattgtgaaaaagtctgtttga"
+	if feature != seq {
+		t.Errorf("Feature GetSequence method has failed. Got this:\n%s instead of \n%s", feature, seq)
+	}
+
+	// Check to see if GetSequence method works on Sequence struct
+	if gbk.GetSequence() != gbk.Sequence {
+		t.Errorf("Sequence GetSequence method has failed.. Got this:\n%s instead of \n%s", gbk.GetSequence(), gbk.Sequence)
+	}
+
+}
+
+func TestLocationParser(t *testing.T) {
+	gbk := ReadGbk("../../data/t4_intron.gb")
+
+	// Read 1..243
+	feature := gbk.Features[1].GetSequence()
+	seq := "atgagattacaacgccagagcatcaaagattcagaagttagaggtaaatggtattttaatatcatcggtaaagattctgaacttgttgaaaaagctgaacatcttttacgtgatatgggatgggaagatgaatgcgatggatgtcctctttatgaagacggagaaagcgcaggattttggatttaccattctgacgtcgagcagtttaaagctgattggaaaattgtgaaaaagtctgtttga"
+	if feature != seq {
+		t.Errorf("Feature sequence parser has changed on test '1..243'. Got this:\n%s instead of \n%s", feature, seq)
+	}
+
+	// Read join(893..1441,2459..2770)
+	featureJoin := gbk.Features[6].GetSequence()
+	seqJoin := "atgaaacaataccaagatttaattaaagacatttttgaaaatggttatgaaaccgatgatcgtacaggcacaggaacaattgctctgttcggatctaaattacgctgggatttaactaaaggttttcctgcggtaacaactaagaagctcgcctggaaagcttgcattgctgagctaatatggtttttatcaggaagcacaaatgtcaatgatttacgattaattcaacacgattcgttaatccaaggcaaaacagtctgggatgaaaattacgaaaatcaagcaaaagatttaggataccatagcggtgaacttggtccaatttatggaaaacagtggcgtgattttggtggtgtagaccaaattatagaagttattgatcgtattaaaaaactgccaaatgataggcgtcaaattgtttctgcatggaatccagctgaacttaaatatatggcattaccgccttgtcatatgttctatcagtttaatgtgcgtaatggctatttggatttgcagtggtatcaacgctcagtagatgttttcttgggtctaccgtttaatattgcgtcatatgctacgttagttcatattgtagctaagatgtgtaatcttattccaggggatttgatattttctggtggtaatactcatatctatatgaatcacgtagaacaatgtaaagaaattttgaggcgtgaacctaaagagctttgtgagctggtaataagtggtctaccttataaattccgatatctttctactaaagaacaattaaaatatgttcttaaacttaggcctaaagatttcgttcttaacaactatgtatcacaccctcctattaaaggaaagatggcggtgtaa"
+	if featureJoin != seqJoin {
+		t.Errorf("Feature sequence parser has changed on test 'join(893..1441,2459..2770)'. Got this:\n%s instead of \n%s", featureJoin, seqJoin)
+	}
+
+	// Read complement(2791..3054)
+	featureComplement := gbk.Features[10].GetSequence()
+	seqComplement := "ttattcactacccggcatagacggcccacgctggaataattcgtcatattgtttttccgttaaaacagtaatatcgtagtaacagtcagaagaagttttaactgtggaaattttattatcaaaatactcacgagtcattttatgagtatagtattttttaccataaatggtaataggctgttctggtcctggaacttctaactcgcttgggttaggaagtgtaaaaagaactacaccagaagtatctttaaatcgtaaaatcat"
+	if featureComplement != seqComplement {
+		t.Errorf("Feature sequence parser has changed on test 'complement(2791..3054)'. Got this:\n%s instead of \n%s", featureComplement, seqComplement)
+	}
+
+	// Read join(complement(315..330),complement(339..896))
+	// Note: it is known that some software, like Snapgene, assumes that since both strands are in the reverse direction
+	// that the first sequence should be appended to the reverse sequence, instead of the second sequence
+	// getting appended to the first. Biopython appends the second sequence to the first, and that is logically
+	// the most obvious thing to do, so we are implementing it that way.
+	featureJoinComplement := gbk.Features[3].GetSequence()
+	seqJoinComplement := "ataccaatttaatcattcatttatatactgattccgtaagggttgttacttcatctattttataccaatgcgtttcaaccatttcacgcttgcttatatcatcaagaaaacttgcgtctaattgaactgttgaattaacacgatgccttttaacgatgcgagaaacaactacttcatctgcataaggtaatgcagcatataacagagcaggcccgccaattacacttactttagaattctgatcaagcatagtttcgaatggtgcattagggcttgacacttgaatttcgccgccagaaatgtaagttatatattgctcccaagtaatatagaaatgtgctaaatcgccgtctttagttacaggataatcacgcgcaaggtcacacaccacaatatggctacgaccaggaagtaatgtaggcaatgactggaacgttttagcacccataatcataattgtgccttcagtacgagctttaaaattctggaggtcctttttaactcgtccccatggtaaaccatcacctaaaccgaatgctaattcattaaagccgtcgaccgttttagttggaga"
+	if featureJoinComplement != seqJoinComplement {
+		t.Errorf("Feature sequence parser has changed on test 'join(complement(315..330),complement(339..896))'. Got this:\n%s instead of \n%s", featureJoinComplement, seqJoinComplement)
+	}
+
+	// Read complement(join(893..1098,1101..2770))
+	featureComplementJoin := gbk.Features[5].GetSequence()
+	seqComplementJoin := "ttacaccgccatctttcctttaataggagggtgtgatacatagttgttaagaacgaaatctttaggcctaagtttaagaacatattttaattgttctttagtagaaagatatcggaatttataaggtagaccacttattaccagctcacaaagctctttaggttcacgcctcaaaatttctttacattgttctacgtgattcatatagatatgagtattaccaccagaaaatatcaaatcccctggaataagattacacatcttagctacaatatgaactaacgtagcatatgacgcaatattaaacggtagcattatgttcagataaggtcgttaatcttaccccggaattatatccagctgcatgtcaccatgcagagcagactatatctccaacttgttaaagcaagttgtctatcgtttcgagtcacttgaccctactccccaaagggatagtcgttaggcatttatgtagaaccaattccatttatcagattttacacgataagtaactaatccagacgaaattttaaaatgtctagctgcatctgctgcacaatcaaaaataaccccatcacatgaaatctttttaatattactaggctttttacctttcatcttttctgatattttagatttagttatgtctgaatgcttatgattaaagaatgaattattttcacctgaacgatttctgcatttactacaagtataagcagaagtttgtatgcgaacaccgcacttacaaaacttatgggtttctggattccaacgcccgtttttacttccgggtttactgtaaagagctttccgaccatcaggtccaagtttaagcatcttagctttaacagtttcagaacgtttcttaataatttcttcttttaatggatgcgtagaacatgtatcaccaaacgttgcatcagcaatattgtatccattaattttagaattaagctctttaatccaaaaattttctcgttcaataatcaaatctttctcatatggaatttcttccaaaatagaacattcaaacacattaccatgtttgttaaaagacctctgaagttttatagaagaatggcatcctttttctaaatctttaaaatgcctcttccatctcttttcaaaatctttagcacttcctacatatactttattgtttaaagtatttttaatctgataaattccgcttttcataaatacctctttaaatatagaagtatttattaaagggcaagtcctacaatttagcacgggattgtctactagagaggttccccgtttagatagattacaagtataagtcaccttatactcaggcctcaattaacccaagaaaacatctactgagcgttgataccactgcaaatccaaatagccattacgcacattaaactgatagaacatatgacaaggcggtaatgccatatatttaagttcagctggattccatgcagaaacaatttgacgcctatcatttggcagttttttaatacgatcaataacttctataatttggtctacaccaccaaaatcacgccactgttttccataaattggaccaagttcaccgctatggtatcctaaatcttttgcttgattttcgtaattttcatcccagactgttttgccttggattaacgaatcgtgttgaattaatcgtaaatcatacatttgtgcttcctgataaaaaccatattagctcagcaatgcaagctttccaggcgagcttcttagttgttaccgcaggaaaacctttagttaaatcccagcgtaatttagatccgaacagagcaattgttcctgtgcctgtacgatcatcggtttcataaccattttcaaaaatgtctttaattaaatcttggtattgtttcat"
+	if featureComplementJoin != seqComplementJoin {
+		t.Errorf("Feature sequence parser has changed on test 'complement(join(893..1098,1101..2770))'. Got this:\n%s instead of \n%s", featureComplementJoin, seqComplementJoin)
+	}
+}
+
+func TestGenbankNewlineParsingRegression(t *testing.T) {
+	gbk := ReadGbk("../../data/puc19.gbk")
+
+	for _, feature := range gbk.Features {
+		if feature.SequenceLocation.Start == 410 && feature.SequenceLocation.End == 1750 && feature.Type == "CDS" {
+			if feature.Attributes["product"] != "chromosomal replication initiator informational ATPase" {
+				t.Errorf("Newline parsing has failed.")
+			}
+			break
+		}
+	}
+}
+
+func BenchmarkReadGbk(b *testing.B) {
+	for i := 0; i < b.N; i++ {
+		ReadGbk("../../data/bsub.gbk")
+	}
+}
+
+func BenchmarkReadGbk1(b *testing.B)     { BenchmarkReadGbk(b) }
+func BenchmarkReadGbk10(b *testing.B)    { BenchmarkReadGbk(b) }
+func BenchmarkReadGbk100(b *testing.B)   { BenchmarkReadGbk(b) }
+func BenchmarkReadGbk1000(b *testing.B)  { BenchmarkReadGbk(b) }
+func BenchmarkReadGbk10000(b *testing.B) { BenchmarkReadGbk(b) }
+
+/******************************************************************************
+
+Gbk/gb/genbank related benchmarks end here.
+
+******************************************************************************/
+
+/******************************************************************************
+
+GbkMulti/GbkFlat related tests begin here.
+
+******************************************************************************/
+
+func ExampleReadGbkMulti() {
+	sequences := ReadGbkMulti("../../data/multiGbk_test.seq")
+	var locus []string
+	for _, sequence := range sequences {
+		locus = append(locus, sequence.Meta.Locus.Name)
+	}
+
+	fmt.Println(strings.Join(locus, ", "))
+	// Output: AB000100, AB000106
+}
+
+func ExampleReadGbkFlat() {
+	sequences := ReadGbkFlat("../../data/long_comment.seq")
+	var locus []string
+	for _, sequence := range sequences {
+		locus = append(locus, sequence.Meta.Locus.Name)
+	}
+
+	fmt.Println(strings.Join(locus, ", "))
+	// Output: AB000100, AB000106
+}
+
+func ExampleReadGbkFlatGz() {
+	sequences := ReadGbkFlatGz("../../data/flatGbk_test.seq.gz")
+	//sequences := ReadGbkFlatGz("../../data/gbbct358.seq.gz")
+	var locus []string
+	for _, sequence := range sequences {
+		locus = append(locus, sequence.Meta.Locus.Name)
+	}
+	fmt.Println(strings.Join(locus, ", "))
+	// Output: AB000100, AB000106
+}
+
+func ExampleParseGbkMulti() {
+	file, _ := ioutil.ReadFile("../../data/multiGbk_test.seq")
+	sequences := ParseGbkMulti(file)
+	var locus []string
+	for _, sequence := range sequences {
+		locus = append(locus, sequence.Meta.Locus.Name)
+	}
+
+	fmt.Println(strings.Join(locus, ", "))
+	// Output: AB000100, AB000106
+}
+
+func ExampleParseGbkFlat() {
+	file, _ := ioutil.ReadFile("../../data/flatGbk_test.seq")
+	sequences := ParseGbkFlat(file)
+	var locus []string
+	for _, sequence := range sequences {
+		locus = append(locus, sequence.Meta.Locus.Name)
+	}
+
+	fmt.Println(strings.Join(locus, ", "))
+	// Output: AB000100, AB000106
+}
+
+/******************************************************************************
+
+GbkMulti/GbkFlat related tests end here.
+
+******************************************************************************/
