+++ io/uniprot/data/xml_go_generator
