+++ io/uniprot/uniprot.go
