+++ transformations/transformations_test.go
@@ -1,14 +1,46 @@
-package poly
+package transformations
 
 import (
 	"fmt"
 	"io/ioutil"
+	"log"
 	"os"
+	"sort"
 	"strings"
 	"testing"
 
+	"github.com/TimothyStiles/poly/io/genbank"
 	"github.com/google/go-cmp/cmp"
 )
+
+func TestIUPAC(t *testing.T) {
+	testSeq := "ATN"
+	testVariants := []string{"ATG", "ATA", "ATT", "ATC"}
+	testVariantsIUPAC, err := AllVariantsIUPAC(testSeq)
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	sort.Strings(testVariants)
+	sort.Strings(testVariantsIUPAC)
+
+	for index := range testVariants {
+		if testVariants[index] != testVariantsIUPAC[index] {
+			t.Errorf("IUPAC variant has changed on test 'allIUPAC('ATN')'. Got this:\n%s instead of \n%s", testVariantsIUPAC, testVariants)
+		}
+	}
+}
+
+func ExampleAllVariantsIUPAC() {
+	// AllVariantsIUPAC takes a string as input
+	// and returns all iupac variants as output
+	mendelIUPAC := "ATGGARAAYGAYGARCTN"
+	// ambiguous IUPAC codes for most of the sequences that code for the protein MENDEL
+	mendelIUPACvariants, _ := AllVariantsIUPAC(mendelIUPAC)
+	fmt.Println(mendelIUPACvariants)
+	// Output: [ATGGAGAATGATGAGCTG ATGGAGAATGATGAGCTA ATGGAGAATGATGAGCTT ATGGAGAATGATGAGCTC ATGGAGAATGATGAACTG ATGGAGAATGATGAACTA ATGGAGAATGATGAACTT ATGGAGAATGATGAACTC ATGGAGAATGACGAGCTG ATGGAGAATGACGAGCTA ATGGAGAATGACGAGCTT ATGGAGAATGACGAGCTC ATGGAGAATGACGAACTG ATGGAGAATGACGAACTA ATGGAGAATGACGAACTT ATGGAGAATGACGAACTC ATGGAGAACGATGAGCTG ATGGAGAACGATGAGCTA ATGGAGAACGATGAGCTT ATGGAGAACGATGAGCTC ATGGAGAACGATGAACTG ATGGAGAACGATGAACTA ATGGAGAACGATGAACTT ATGGAGAACGATGAACTC ATGGAGAACGACGAGCTG ATGGAGAACGACGAGCTA ATGGAGAACGACGAGCTT ATGGAGAACGACGAGCTC ATGGAGAACGACGAACTG ATGGAGAACGACGAACTA ATGGAGAACGACGAACTT ATGGAGAACGACGAACTC ATGGAAAATGATGAGCTG ATGGAAAATGATGAGCTA ATGGAAAATGATGAGCTT ATGGAAAATGATGAGCTC ATGGAAAATGATGAACTG ATGGAAAATGATGAACTA ATGGAAAATGATGAACTT ATGGAAAATGATGAACTC ATGGAAAATGACGAGCTG ATGGAAAATGACGAGCTA ATGGAAAATGACGAGCTT ATGGAAAATGACGAGCTC ATGGAAAATGACGAACTG ATGGAAAATGACGAACTA ATGGAAAATGACGAACTT ATGGAAAATGACGAACTC ATGGAAAACGATGAGCTG ATGGAAAACGATGAGCTA ATGGAAAACGATGAGCTT ATGGAAAACGATGAGCTC ATGGAAAACGATGAACTG ATGGAAAACGATGAACTA ATGGAAAACGATGAACTT ATGGAAAACGATGAACTC ATGGAAAACGACGAGCTG ATGGAAAACGACGAGCTA ATGGAAAACGACGAGCTT ATGGAAAACGACGAGCTC ATGGAAAACGACGAACTG ATGGAAAACGACGAACTA ATGGAAAACGACGAACTT ATGGAAAACGACGAACTC]
+
+}
 
 func ExampleTranslate() {
 
@@ -69,7 +101,7 @@ func ExampleOptimize() {
 
 	gfpTranslation := "MASKGEELFTGVVPILVELDGDVNGHKFSVSGEGEGDATYGKLTLKFICTTGKLPVPWPTLVTTFSYGVQCFSRYPDHMKRHDFFKSAMPEGYVQERTISFKDDGNYKTRAEVKFEGDTLVNRIELKGIDFKEDGNILGHKLEYNYNSHNVYITADKQKNGIKANFKIRHNIEDGSVQLADHYQQNTPIGDGPVLLPDNHYLSTQSALSKDPNEKRDHMVLLEFVTAAGITHGMDELYK*"
 
-	sequence := ReadGbk("data/puc19.gbk")
+	sequence := genbank.ReadGbk("data/puc19.gbk")
 	codonTable := GetCodonTable(11)
 
 	optimizationTable := sequence.GetOptimizationTable(codonTable)
